+++ CHANGELOG.md
@@ -9,3 +9,4 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Initial release
 - Add one click package upload function
 - Add readme.md Chinese English switching and fix English Env variable obtain
+- Add create mobile project template